allow interpolation of SCM urls
ammend #442 to allow use of project.artifactId and artifactId as
properties inside the SCM section

testing a project with maven 4.0.0-alpha4 showed no warnings or errors
about this use case, and we have many plugins that use it.

Whilst fixing all the plugin is nobale - that would unessescarily hold
up future improvements.