Native support for multi-module Maven projects (#510)
This changes the way the PCT clones from per plugin to per repository.

Once the repository is cloned it then exercises each plugin in the repo in turn (serially).
plugins that are excluded (or implicitly if an include list is provided) will not be tested even if some other plugin in the same repository is tested. It is assumed that all plugins in the same repository will use the same tag (however this is not explicitly prevented - but can be if deemed necessary)

The pct command is no longer a root command but is now a sub command test-plugins and any caller of the PCT will need to be adjusted to use this sub command.

There is a new command list-plugins that will list all plugins in the war (excluding detached) that can also filter by includes or excludes.

Tooling can call this in order to group plugin tests for the same repository. (as above this does not group by repository and tag - it is assumed that the version of plugins from the same repository will be the same - this can be added if required, or we can group by repository and commit)

All AbstractMultiParentHook have been removed and similar to the git url workaround has been inlined into code that should become dead once all plugins have been updated to include jenkinsci/maven-hpi-plugin#436
The support is currently via a hook like MetadataExtractor - all implkementations of this are expected to be removed when plugins are migrated, simplifying the implementation.

---------

Co-authored-by: Basil Crow <me@basilcrow.com>