re-introduce check on tag alignment for a repository
In the multimodule branch there was a check to ensure that all plugins
from a given repository use the same tag, but this was not carried over
in #510

this changes the assumption that the input is correct and could silently
provide invorrect results, to an explicit check that fails early before
starting to check compatability.

Only plugins that are being checked for compatability will be involved
in the repository.  This is to allows a war to contain a plugin that was
formerly in the repository that has been changed to just migrate data to
a new location (by adding a dependency on a replacement plugin) to still
be in the war, even after it has been deleted from the repository.