Support testing SNAPSHOTs that have a deterministic git revision
If a SNAPSHOT build is in the override war and it was built with a
version of maven-hpi-plugin 3.42 or higher (maven-hpi-plugin#436) then
we have a concrete git sha we can use for checkout of the plugin.

So this relaxes the enforcement of not building non releases by allowing
to test a SNAPSHOT also so long as we have a git commit for it.

in the case the commit is HEAD (which is set by the legacy plugin
extractors) then we will refuse to test the plugin the same as we do
today.