Adapt to authentication-tokens 1.4 git:// in pom file
https://github.com/jenkinsci/bom/issues/2095 requests the addition of
authentication tokens to the managed set in the plugin bill of materials.

https://github.com/jenkinsci/bom/pull/2189 implements that but shows
that the addition fails because authentication-tokens plugin 1.4 needs
to be pinned for Jenkins 2.375.x and authentication-tokens plugin 1.4
lists its SCM location with a git:// protocol URL instead of an https://
protocol URL.

Use the same workaround that is used for other older plugins.

This can be removed once the 2.375.x line is no longer needed in the
plugin bill of materials.